restructure scrapers, don't run jpl task on dev
